Make the image loading asynchronous.
Now that we are also loading avatars there is a lot of image loading to do to show the timeline.
We can do the loading asynchronously to let the user have a look at the toots already while image loading is incrementally proceeding.

We can no longer enforce caching of avatar loading since the variable is consulted when the response parsing happens at which point the dynamic binding we had used so far has gone out of scope again.