Catch any errors thrown during url-retrieve.
On Emacs24 I've been able to reliably fail url fetching which exposed issues in loading a timeline (it aborts the loading).
This catches any errors, marking the image load as failed so that we won't retry (retries are a TODO item I guess) and then succeeds the function so the rest of the timeline loading can proceed.