Closes #152 and extends the fix for #150 as well as a host of bug fixes
We now kill the http get request buffer once JSON has been extracted.

mastodon-tl--as-string was implemented and replaced any occurrence of number-to-string or int-to-string

Added variable mastodon-tl--display-media-p. By default it is 't but can be made a local buffer variable and set to nil. When nil rather than displaying the media it just provides a link Media::<link>

Fixed checking for faves and boosts, they should now render properly. The return from json-read-from-string for nil is :json-false which evaluates to 't in elisp.

Fixed the format string that gets printed when faving and boosting

Fixed mastodon-tl--thread updating and requesting and changed its behaviour such that it tries to open the original toot thread rather than the boosted thread.

Added tests for both the new mastodon-tl--as-string function and the mastodon-tl--toot-id utility.

enter mastodon mode before defining local buffer variable mastodon-tl--buffer-spec. This fixes some oddities with the local buffer variable.