Make "Content warning" a tab stop and toggle to show/hide the main contents. (#170)
* Allow user to navigate interesting things in a buffer via tabbing (tab to go forward, M-tab and S-Tab to go back).

This has always been possible while on a hyperlink but now works everywhere.
Currently only hyperlinks are tab stops but in the future we will want to support other things and there are already TODO comments in the code to note where we may want to do this.

* Add a new tab stop and link type: spoiler toggling.

This initially hides the spoiler in a toot and  makes the "Content warning" a link and tab stop. The action taken is to toggle the visibility of the toot.