Fix the boosting/favouriting of statuses in profiles.
(Favouriting itself is still hard since we have clobbered the 'f' keybinding, but if you bind it to something else or invoke it via `M-x mastodon-toot--toggle-favourite` then at least it works.)
This also changes the regular boosting/favoriting behavior.  Before we would boost/fav a boost or fav instead of the actual toot that was boosted/faved.  With this change we always boost/fav the base toot.

To do this we now keep a second toot id (with the base toot) in the byline's properities.  (For regular statuses 'toot-id and 'base-toot-id will be identical.)