Client id, secret are not to be set explicitly by the user
`mastodon-client--register` - which is an existing method - handles
the generation of new application in the user's account, store the ID
and token locally. The simplest entry point to read/fetch the
application details is to call `mastodon-client` because it reads if
the values are locally available, or makes appropriate API calls to
fetch them from the instance; making the life simple by returning an
easily accessible plist.

Thus, I have removed the requirement of user having to create new
application themselves. Now the `mastodon-client` is called while
generating new oauth2 tokens.