feed notfication ID to mastodon-notifications--insert-status
which is a copy of mastodon-tl--insert-status. this makes the 'toot-id of notifications
that of the favoriting/boosting item, rather than the item boosted/favorited.

this ID is needed in order to make loading more older notifications work correctly.
beforehand, the call would be incorrect and the same latest 20 notifications would load.