merge upload-media-attachments functionality into toot-send.
this obviates the need for the user to upload files before posting their toot.

- this commit makes http--post-media-attachment synchronous, so that toot-send
has to wait for it.

- in toot-send: if mastodon-toot--media-attachements is non-nil, the files it
contains are uploaded synchronously, and their returned ids are added to
toot-media-attachment-ids, which are parsed as args for the POST request to
be attached to the toot.

- then we send toot as usual.

- clear-all-attachments also clears mastodon-toot--media-attachment-ids just
in case.

- we have no more need of media-attachments-filenames, as media-attachments is
now a list and not a boolean value.