Use portable filename component functions.
Apparently one should not rely on "/" being the directory separator
and use the funtions from
https://www.gnu.org/software/emacs/manual/html_node/elisp/File-Name-Components.html#File-Name-Components
instead.

The new version seems strictly better in that it won't create paths
with double slashes when `emojify-emojis-dir` already ends in a slash.

This also refines the test for `emojify-emojis-dir` to actually check
it is an existing directoy and not just an existing file, dir, or
symlink.