FIX echo faves in thread view clashing with media help-echo
move help-echo propertizing to author name of byline, in
`mastodon-tl--byline-author'. so when we use `goto-next-toot' point is on
author and we get the help echo

and add the same help-echo to `mastodon-tl--content', but only when in thread
view. this ensures the propertizing occurs prior to any image propertizing,
when will then only replace the faves propertizing for the image part of the
given toot.

that way we have help echo for images when we want, and faves count otherwise.