refactor mastodon-notifications--insert-status
which was a copy of mastodon-tl--insert-status.

we revert to having just the latter as main function with optional argument.

mastodon-notifications--insert-status just calls it with the arg ID.

the reason we need the difference is to ensure notifications have their own
ID, and not that of the toot the notif refers to, attached as property
"toot-id". then we have all functionality working on notifications, such as
boosting mentions and so on.